backend/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
boto3==1.14.16
botocore==1.17.16
cachetools==4.0.0
chardet==3.0.4
coverage==5.0.2
Django==3.0.7
django-csp==3.7
django-extensions==2.2.6
django-filter==2.3.0
django-sslserver==0.22
django-storages==1.9.1
django-webtest==1.9.7
djangorestframework==3.11.0
docutils==0.15.2
google-auth==1.10.0
idna==2.8
jmespath==0.10.0
markdown==3.2.2
moviepy==1.0.3
Pillow==7.2.0
pipenv==2018.11.26
psycopg2-binary==2.8.4
ptvsd==4.3.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pypugjs==5.8.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
s3transfer==0.3.3
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
uWSGI==2.0.18
waitress==1.4.3
WebOb==1.8.5
WebTest==2.0.33
whitenoise==5.0.1
